% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_outerproduct}
\alias{rcpparma_outerproduct}
\title{rcpparma_outerproduct}
\usage{
rcpparma_outerproduct(x)
}
\arguments{
\item{x}{A vector for which the outer product is calculated}
}
\value{
\item{First}{my first item}
\item{Second}{second element of list}
}
\description{
My silly function
}
\details{
A longer description
}
\examples{
x <- rnorm(10)
rcpparma_outerproduct(x)

}
\author{
Spencer Woody
}
